Move detailview fragments to new folder
